Use babel-preset-es2015
Because we want to use ES6 and translate to ES5, use
`babel-preset-es2015`. We also use `compile` to run our code though
babel and we set it before the tests and before publish to NPM. Although
the output path is /lib, we still want lint to run to our original code
(/src) instead of the translated one.

Also Eslint now:
- Use jest configuration to avoid having to specify its global vars
- Using `console` raises a warning instead of an error.
- sourceType set to module